Update README.md
This is a simple machine learning project that is made with a simple dataset and takes the height and weight from the user, and according to that, the KNN algorithm makes predictions. And if we want to enter the details of the project, this project comes after receiving the inputs from the user and finds 3 of the closest neighbors to the input information or the same user by Euclidean distance. The number of each is more, it will be printed in the output